requirements.txt
email-validator
redis
hiredis
flake8
fastapi==0.65.1
uvicorn[standard]~=0.13.4
pydantic~=1.8.1
simplejson~=3.17.2
aiohttp~=3.7.4.post0
Pillow~=8.2.0
deeppyer~=1.0.1
discord.py
